PDFBOX-2147: Removed unnecessary unboxing
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1613617 13f79535-47bb-0310-9956-ffa450edef68

+++ fontbox/src/main/java/org/apache/fontbox/cff/Type1CharStringFormatter.java
@@ -67,7 +67,7 @@ public class Type1CharStringFormatter
 
     private void writeNumber(Integer number)
     {
-        int value = number.intValue();
+        int value = number;
         if (value >= -107 && value <= 107)
         {
             output.write(value + 139);
